+++ frontend/src/views/settings/Profile.vue
@@ -23,10 +23,7 @@
               v-model="localuser.quickDownload"
               :name="$t('profileSettings.showQuickDownload')"
             />
-          </div>
-          <h3>{{ $t("profileSettings.filePreviewOptions") }}</h3>
-          <div class="settings-items">
-            <ToggleSwitch
+                        <ToggleSwitch
               class="item"
               v-model="localuser.preview.image"
               :name="$t('profileSettings.previewImages')"
@@ -58,6 +55,19 @@
               class="item"
               v-model="localuser.preview.popup"
               :name="$t('profileSettings.popupPreview')"
+            />
+          </div>
+          <h3>{{ $t("profileSettings.filePreviewOptions") }}</h3>
+          <div class="settings-items">
+            <ToggleSwitch
+              class="item"
+              v-model="localuser.preview.autoplayMedia"
+              :name="$t('profileSettings.autoplayMedia')"
+            />
+            <ToggleSwitch
+              class="item"
+              v-model="localuser.preview.disableHideSidebar"
+              :name="$t('profileSettings.disableHideSidebar')"
             />
           </div>
           <div v-if="hasOnlyOfficeEnabled">
@@ -107,12 +117,7 @@
               </button>
             </div>
           </div>
-          <ToggleSwitch
-            v-if="localuser.permissions?.admin"
-            class="item"
-            v-model="localuser.disableUpdateNotifications"
-            :name="$t('profileSettings.disableUpdateNotifications')"
-          />
+  
           <h3>{{ $t("settings.themeColor") }}</h3>
           <ButtonGroup
             :buttons="colorChoices"
@@ -125,6 +130,15 @@
             :locale="localuser.locale"
             @update:locale="updateLocale"
           ></Languages>
+          <h3 v-if="user.permissions.admin">{{ $t("settings.adminOptions") }}</h3>
+          <div v-if="user.permissions.admin" class="settings-items">
+            <ToggleSwitch
+                v-if="localuser.permissions?.admin"
+                class="item"
+                v-model="localuser.disableUpdateNotifications"
+                :name="$t('profileSettings.disableUpdateNotifications')"
+              />
+          </div>
         </div>
       </form>
     </div>
@@ -175,6 +189,9 @@ export default {
     },
   },
   computed: {
+    user() {
+      return state.user;
+    },
     muPdfAvailable() {
       return muPdfAvailable;
     },
@@ -266,5 +283,8 @@ export default {
 <style scoped>
 .card-content h3 {
   text-align: center;
+}
+#officePreviewExt {
+  width: 80%;
 }
 </style>