+++ frontend/tests-playwright/general/file-actions.spec.ts
@@ -76,10 +76,8 @@ test("2x copy from listing to new folder", async({ page, checkForErrors, context
   await page.locator('li[aria-label="myfolder"]').click();
   await expect(page.locator('li[aria-selected="true"]')).toHaveCount(1);
   await page.locator('button[aria-label="Copy"]').click();
-  const popup = page.locator('#popup-notification-content');
-  await popup.waitFor({ state: 'visible' });
-  await expect(popup).toHaveText("Successfully copied file/folder, redirecting...");
-  await page.waitForURL('**/myfolder/');
+  await expect(page.locator('#popup-notification-content')).toHaveText("Resources copied successfully");
+  await page.goto("/files/playwright-files/myfolder/");
   await expect(page).toHaveTitle("Graham's Filebrowser - Files - myfolder");
   // verify exists and copy again
   await page.locator('a[aria-label="copyme.txt"]').waitFor({ state: 'visible' });
@@ -93,9 +91,9 @@ test("2x copy from listing to new folder", async({ page, checkForErrors, context
   await page.locator('input[aria-label="New Folder Name"]').fill('newfolder');
   await page.locator('button[aria-label="Create"]').click();
 
-  await expect(page).toHaveTitle("Graham's Filebrowser - Files - newfolder");
+  await expect(page).toHaveTitle(/.* - newfolder/);
   await page.goBack();
-  await expect(page).toHaveTitle("Graham's Filebrowser - Files - myfolder");
+  await expect(page).toHaveTitle(/.* - myfolder/);
 
   await page.locator('a[aria-label="copyme.txt"]').click( { button: "right" });
   await page.locator('.selected-count-header').waitFor({ state: 'visible' });
@@ -104,11 +102,9 @@ test("2x copy from listing to new folder", async({ page, checkForErrors, context
   await expect(page.locator('div[aria-label="filelist-path"]')).toHaveText('Path: /myfolder/');
   await page.locator('li[aria-label="newfolder"]').click();
   await page.locator('button[aria-label="Copy"]').click();
-  const popup2 = page.locator('#popup-notification-content');
-  await popup2.waitFor({ state: 'visible' });
-  await expect(popup2).toHaveText("Successfully copied file/folder, redirecting...");
-  //await page.waitForURL('**/newfolder/');
-  //await expect(page).toHaveTitle("Graham's Filebrowser - Files - newfolder");
+  await expect(page.locator('#popup-notification-content')).toHaveText("Resources copied successfully");
+  await page.goto("/files/playwright-files/myfolder/newfolder/");
+  await expect(page).toHaveTitle(/.* - newfolder/);
   checkForErrors();
 })
 
@@ -127,7 +123,6 @@ test("delete file", async({ page, checkForErrors, context }) => {
   await popup.waitFor({ state: 'visible' });
   await expect(popup).toHaveText("Deleted item successfully! reloading...");
   checkForErrors();
-
 })
 
 test("delete nested file prompt", async({ page, checkForErrors, context }) => {