+++ frontend/tests-playwright/noauth/file-actions.spec.ts
@@ -52,10 +52,8 @@ test("2x copy from listing to new folder", async({ page, checkForErrors, context
   await page.locator('li[aria-label="myfolder"]').click();
   await expect(page.locator('li[aria-selected="true"]')).toHaveCount(1);
   await page.locator('button[aria-label="Copy"]').click();
-  const popup = page.locator('#popup-notification-content');
-  await popup.waitFor({ state: 'visible' });
-  await expect(popup).toHaveText("Successfully copied file/folder, redirecting...");
-  await page.waitForURL('**/myfolder/');
+  await expect(page.locator('#popup-notification-content')).toHaveText("Resources copied successfully");
+  await page.goto("/files/myfolder/");
   await expect(page).toHaveTitle("Graham's Filebrowser - Files - myfolder");
   // verify exists and copy again
   await page.locator('a[aria-label="copyme.txt"]').waitFor({ state: 'visible' });
@@ -69,9 +67,9 @@ test("2x copy from listing to new folder", async({ page, checkForErrors, context
   await page.locator('input[aria-label="New Folder Name"]').fill('newfolder');
   await page.locator('button[aria-label="Create"]').click();
 
-  await expect(page).toHaveTitle("Graham's Filebrowser - Files - newfolder");
+  await expect(page).toHaveTitle(/.* - newfolder/);
   await page.goBack();
-  await expect(page).toHaveTitle("Graham's Filebrowser - Files - myfolder");
+  await expect(page).toHaveTitle(/.* - myfolder/);
 
   await page.locator('a[aria-label="copyme.txt"]').click( { button: "right" });
   await page.locator('.selected-count-header').waitFor({ state: 'visible' });
@@ -80,11 +78,9 @@ test("2x copy from listing to new folder", async({ page, checkForErrors, context
   await expect(page.locator('div[aria-label="filelist-path"]')).toHaveText('Path: /myfolder/');
   await page.locator('li[aria-label="newfolder"]').click();
   await page.locator('button[aria-label="Copy"]').click();
-  const popup2 = page.locator('#popup-notification-content');
-  await popup2.waitFor({ state: 'visible' });
-  await expect(popup2).toHaveText("Successfully copied file/folder, redirecting...");
-  await page.waitForURL('**/newfolder/');
-  await expect(page).toHaveTitle("Graham's Filebrowser - Files - newfolder");
+  await expect(page.locator('#popup-notification-content')).toHaveText("Resources copied successfully");
+  await page.goto("/files/myfolder/newfolder/");
+  await expect(page).toHaveTitle(/.* - newfolder/);
   checkForErrors();
 })